Automatically fill the Album Artist tag
This script checks for the values of the "Artist" tag and the "Album Artist" tag. If the "Album Artist" tag is empty or null but the "Artist" tag exists and is the same for all files, it populates the "Album Artist" tag with the value from the "Artist" tag.

This is not a user tag, but a proper use of uncommon Album Artist tag.