Automatically fill the Disc Number tag from Album title
This script checks for the values of the "Album" tag and the "Disc Number" tag. If the "Album" tag contains "CD#" or "CD #" where # is a number between 1-99, it populates the "Disc Number" tag with that number if the "Disc Number" tag is null or empty.

This was built to ideally work with Album tags already scraped and populated via MusicBrainz.